refactor(HistogramSelector): Move scoring interface into Score.js
Separate the scoring interface for better clarity in the main HistogramSelector method. Scoring is
directly tied to HistogramSelector specifics, and uses publicAPI and model.
